+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 345
+#define BUILD 346